classification/requirements.txt
argparse
pyyaml 
torch
scikit-learn
tqdm
numpy
opencv-python
efficientnet_pytorch
Pillow
torchvision
pandas
timm
ftfy
regex